Group imports (KSG Haskell 1.3.0 - 3.5). Refs #14.
Imports should be grouped between external imports, external imports
from the same project (but different libraries) and internal imports.
When groups are used, comments must be used to indicate the nature of
each group. If there are no distinct groups, they are considered to be
all in the same group.